Reorganize layout templates
Originally "footer.html" was included by layout.html within the footer
block, which meant that any page deriving from layout.html needed to
include it within their footer block if they override it. For some
reason, lots of pages override it. This means the the JS is lost and the
menus don't work on mobile. Automatically include the JS in base.html.
Add a body_js block to include JS at the end of the body.