shiso: Add subscription detection
Adds the Subscription table, and necessary columns to link it to
Transaction table. Further, add code which will match newly created (or
modified) transactions to their corresponding subscription. Finally, add
the logic code for searching for new subscriptions.

There is no UI for subscriptions yet, but the underlying code works, and
I'll be able to add it with this framework.