speedtest: updates to support IPv6
IPv6 was busted on my internet! Gotta test both. We need to make some
schema changes to the speed and IP fetching schemas to support
incomplete data, and set timeouts for the IP fetching.

Since the ping implementation is hand-rolled, it's a bit more involved.
The main item was my discovery that the "0 or more" bytes of data in a
ping request are not optional for some network stacks. But also, I
needed to support different ICMP codes, etc.

Together, this should make the speedtest jobs a bit more resilient to
timeout failures, and able to capture better data regarding IPv6. The
speedtest job needs some adjustment as well, in a subsequent commit.

Signed-off-by: Stephen Brennan <stephen@brennan.io>