Ignore transactions in the "Transfer" category
Transfer is a category that means the money moved from one account to
another. Normally, this should simply be a wash since you'll see
opposite transactions in each account. (EG a credit card payment in the
checking, and the corresponding payment received in the credit card
account).

However, there seem to be cases where that doesn't work. EG Venmo
doesn't show the transfers into or out of your balance: just the
payments you make. So for those cases, the transactions won't be a wash.
Further, if you're just looking at spending in a single account, the
Transfer payment could be a significant portion of the payments. I don't
like that, so I want to ignore that category by default. But I do want
it to be something I can toggle, so add an option.

Signed-off-by: Stephen Brennan <stephen@brennan.io>