shiso: add a scheduled sync option
This runs the global sync every morning, and emails if any updates or
errors occur. To do so, we need to configure an email server, and setup
SERVER_NAME and PREFERRED_URL_SCHEME to work outside of a request
context. We also add a reusable utility for sending HTML & text email.

Signed-off-by: Stephen Brennan <stephen@brennan.io>